Improved data issue when sending from Arduino
Data sent from the Arduino was of different size when received in the program. Later found out that the serial port was open when the app is open and thus it constantly sent data without the program listening. I fixed it by telling the Arduino to send data only when asked for and received data was trimmed using the string.Trim() method.  It works now but code should be tidied up and logic should be changed as the program should listen for Arduino request and not the other way around.